requirements.txt
fastapi
uvicorn[standard]
openai-whisper
python-multipart
torch --index-url https://download.pytorch.org/whl/cu121
torchaudio --index-url https://download.pytorch.org/whl/cu121
torchvision --index-url https://download.pytorch.org/whl/cu121
